requirements.txt
appnope==0.1.3
asttokens==2.4.1
Babel==2.13.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
ghp-import==2.1.0
idna==3.4
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.5.0
jupyter_core==5.4.0
Markdown==3.5
MarkupSafe==2.1.3
material==0.1
matplotlib-inline==0.1.6
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.4.7
mkdocs-material-extensions==1.3
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
paginate==0.5.6
pandas==2.1.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2==2.9.5
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pymdown-extensions==10.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.22
stack-data==0.6.3
tornado==6.3.3
traitlets==5.12.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
watchdog==3.0.0
wcwidth==0.2.8
